debuginfo: extract namespace.rs
